✏️ Fix : subscription 멤버 자료형 변경

+++ business/src/main/java/com/mzbr/business/member/dto/SubScriptionId.java
@@ -3,6 +3,6 @@ package com.mzbr.business.member.dto;
 import java.io.Serializable;
 
 public class SubScriptionId implements Serializable {
-	private int follower;
-	private int followee;
+	private long follower;
+	private long followee;
 }
